move shared pos screens into shared package
